requirements.txt
kaggle
pandas
sklearn
marshmallow_dataclass
pyaml
click
hydra-core
mlflow
boto3
Faker